+++ CHANGELOG.md
@@ -5,11 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.1.11
+## 1.2.0
 
 ### Fixed
 
 - Descriptor groupId should default to org id when provisioning
+
+### Changed
+
+- enpoint is deprecated. Use implementationUrl and/or consumerUrl
+- endpointJson is deprecated, use consumerUrlJson instead
 
 ## 1.1.10